+++ backend/db/data/key_items.go
@@ -74,6 +74,7 @@ var KeyItems = map[uint32]ItemData{
 	0x40002458: {Name: "Fugitive Warrior's Recipe [5]", Category: "key_items", MaxInventory: 1, MaxStorage: 0, MaxUpgrade: 0, IconPath: "items/key_items/fugitive_warriors_recipe_[5].png", Flags: []string{"cut_content", "ban_risk"}},
 	0x40002297: {Name: "Fire Monks' Prayerbook", Category: "key_items", MaxInventory: 1, MaxStorage: 0, MaxUpgrade: 0, IconPath: "items/key_items/fire_monks_prayerbook.png"},
 	0x401EA3CC: {Name: "Cross-Marked Map", Category: "key_items", MaxInventory: 1, MaxStorage: 0, MaxUpgrade: 0, IconPath: "items/key_items/cross_marked_map.png", Flags: []string{"dlc"}},
+	0x401EA3D3: {Name: "Black Syrup", Category: "key_items", MaxInventory: 1, MaxStorage: 0, MaxUpgrade: 0, IconPath: "items/tools/black_syrup.png", Flags: []string{"dlc"}},
 	0x400021D4: {Name: "Mirage Riddle", Category: "key_items", MaxInventory: 1, MaxStorage: 0, MaxUpgrade: 0, IconPath: "items/key_items/mirage_riddle.png"},
 	0x401EA749: {Name: "Greasemonger's Bell Bearing", Category: "key_items", MaxInventory: 1, MaxStorage: 0, MaxUpgrade: 0, IconPath: "items/key_items/greasemongers_bell_bearing.png", Flags: []string{"dlc"}},
 	0x40001FA9: {Name: "Dectus Medallion (Left)", Category: "key_items", MaxInventory: 1, MaxStorage: 0, MaxUpgrade: 0, IconPath: "items/key_items/dectus_medallion_(left).png"},